requirements.txt
dnspython==2.4.2
flask==3.0.0
flask-cors==4.0.0
